feature: add support for cross repo navigation
This change captures "packages" that are then passed into the
`ScipSemanticdbOptions` so that the `index.scip` file as the information
it needs about the dependencies that are being used in the project. We
do persist this file next to the scip file, not next to the semanticdb
like is usually done. The main reason for this right now is that we're
the one putting together the actually scip, so I don't _believe_
anything will actually be using this file.